Functionality fixes and additions.
Fixed image zoom syncronization. The issue hasn't surfaced again as of
yet.

Fixed the mouse direction animation that fades between Before and After
when more than one slider is on the page causing each sliders elements
to trigger at once when only one is being dragged.

Added a mouseup event listener to fire outside the current mousedown
element to reset the cocoen-drag opacity.

Changed the code for multiple Cocoens to be cross browser compatible.
The original .foreach() function was not working in IE.